File structure
Classes were thansfered to their own files
